+++ src/Graphic/DrawLineWinner.c
@@ -7,19 +7,19 @@
 #include "../Test/Test.h"
 #include "../Game/Game.h"
 
-static int_fast8_t signe(const int_fast32_t n);
-static void ligneHorizontale(const Game *game, const int_fast16_t x, const int_fast16_t y, const int_fast16_t w);
-static void disque(Game *game, const int_fast16_t cx, const int_fast16_t cy, const int_fast16_t rayon);
+static int signe(const int n);
+static void ligneHorizontale(const Game *game, const int x, const int y, const int w);
+static void disque(Game *game, const int cx, const int cy, const int rayon);
 
 void lineWinner(Game *game, const Coor dpt, const Coor arv)
 {
-	const int_fast8_t cx = signe(arv.x - dpt.x);
-	const int_fast8_t cy = signe(arv.y - dpt.y);
+	const int cx = signe(arv.x - dpt.x);
+	const int cy = signe(arv.y - dpt.y);
 
-	uint_fast16_t x = dpt.x * (SIZE_X / 3) + SIZE_X / 6;
-	uint_fast16_t y = dpt.y * (SIZE_X / 3) + SIZE_X / 6 + OF_SET;
-	const uint_fast16_t x_arv = arv.x * (SIZE_X / 3) + SIZE_X / 6;
-	const uint_fast16_t y_arv = arv.y * (SIZE_X / 3) + SIZE_X / 6 + OF_SET;
+	int x = dpt.x * (SIZE_X / 3) + SIZE_X / 6;
+	int y = dpt.y * (SIZE_X / 3) + SIZE_X / 6 + OF_SET;
+	const int x_arv = arv.x * (SIZE_X / 3) + SIZE_X / 6;
+	const int y_arv = arv.y * (SIZE_X / 3) + SIZE_X / 6 + OF_SET;
 
 	Test(game->window, game->renderer, WARNING, SDL_SetRenderDrawColor(game->renderer, 77, 218, 225, 50) != 0, "SDL_SetRenderDrawColor", __FILE__, __LINE__);
 	Test(game->window, game->renderer, WARNING, SDL_SetRenderDrawBlendMode(game->renderer, SDL_BLENDMODE_BLEND) != 0, "SDL_SetRenderDrawBlendMode", __FILE__, __LINE__);
@@ -37,9 +37,9 @@ void lineWinner(Game *game, const Coor dpt, const Coor arv)
 	SDL_Delay(500);
 }
 
-static void disque(Game *game, const int_fast16_t cx, const int_fast16_t cy, const int_fast16_t rayon)
+static void disque(Game *game, const int cx, const int cy, const int rayon)
 {
-	int_fast16_t d, y, x;
+	int d, y, x;
 
 	d = 3 - (2 * rayon);
 	x = 0;
@@ -64,7 +64,7 @@ static void disque(Game *game, const int_fast16_t cx, const int_fast16_t cy, con
 	}
 }
 
-static void ligneHorizontale(const Game *game, const int_fast16_t x, const int_fast16_t y, const int_fast16_t w)
+static void ligneHorizontale(const Game *game, const int x, const int y, const int w)
 {
 	SDL_Rect r;
 
@@ -76,7 +76,7 @@ static void ligneHorizontale(const Game *game, const int_fast16_t x, const int_f
 	Test(game->window, game->renderer, WARNING, SDL_RenderFillRect(game->renderer, &r) != 0, "SDL_RenderFillRect", __FILE__, __LINE__);
 }
 
-static int_fast8_t signe(const int_fast32_t n)
+static int signe(const int n)
 {
 	if (n < 0)
 	{
