+++ src/Grille/Grille.h
@@ -3,10 +3,10 @@
 
 #include "../Game/Game.h"
 
-bool grilleVide(int_fast8_t (*grille)[LEN_GRILLE]);
-bool grillePlein(int_fast8_t (*grille)[LEN_GRILLE]);
-void resetGrille(int_fast8_t (*grille)[LEN_GRILLE]);
-void printGrille( Game* game, const uint_fast8_t w);
-bool addGrille( Game* game, const uint_fast16_t clickX, const uint_fast16_t clickY);
+bool grilleVide(int (*grille)[LEN_GRILLE]);
+bool grillePlein(int (*grille)[LEN_GRILLE]);
+void resetGrille(int (*grille)[LEN_GRILLE]);
+void printGrille( Game* game, const int w);
+bool addGrille( Game* game, const int clickX, const int clickY);
 
 #endif // !GRILLE_H_INCLUDED
